Hmm, lets use full last names in display names
The recursive function was clever, but I'm not sure that it was
doing the right thing.

Consider Chris Readinger and Chris Randall. Both show on the site
now with their full names. Which is fine.

It would not be better for them to have been assigned "Chris Re" and
"Chris Ra" when they registered.