Get Docker building again - switch to alpine
The previous pip wheel-based installer had bitrotted due to
changes in pip internals. I could not get it to work quite right
even after updating setup.py, so I switched this to use the offical
Python 3.7 / Alpine container as a base, much as the freezing-web
project does now. That gets this to build. But it will be another thing
to test it, I hope this works.